Split LSP completion.rs into several files
