+++ tooling/lsp/src/requests/completion/builtins.rs
@@ -1,4 +1,64 @@
+use lsp_types::CompletionItemKind;
 use noirc_frontend::token::Keyword;
+use strum::IntoEnumIterator;
+
+use super::{
+    completion_items::{simple_completion_item, snippet_completion_item},
+    name_matches, NodeFinder,
+};
+
+impl<'a> NodeFinder<'a> {
+    pub(super) fn builtin_functions_completion(&mut self, prefix: &str) {
+        for keyword in Keyword::iter() {
+            if let Some(func) = keyword_builtin_function(&keyword) {
+                if name_matches(func.name, prefix) {
+                    self.completion_items.push(snippet_completion_item(
+                        format!("{}(…)", func.name),
+                        CompletionItemKind::FUNCTION,
+                        format!("{}({})", func.name, func.parameters),
+                        Some(func.description.to_string()),
+                    ));
+                }
+            }
+        }
+    }
+
+    pub(super) fn builtin_values_completion(&mut self, prefix: &str) {
+        for keyword in ["false", "true"] {
+            if name_matches(keyword, prefix) {
+                self.completion_items.push(simple_completion_item(
+                    keyword,
+                    CompletionItemKind::KEYWORD,
+                    Some("bool".to_string()),
+                ));
+            }
+        }
+    }
+
+    pub(super) fn builtin_types_completion(&mut self, prefix: &str) {
+        for keyword in Keyword::iter() {
+            if let Some(typ) = keyword_builtin_type(&keyword) {
+                if name_matches(typ, prefix) {
+                    self.completion_items.push(simple_completion_item(
+                        typ,
+                        CompletionItemKind::STRUCT,
+                        Some(typ.to_string()),
+                    ));
+                }
+            }
+        }
+
+        for typ in builtin_integer_types() {
+            if name_matches(typ, prefix) {
+                self.completion_items.push(simple_completion_item(
+                    typ,
+                    CompletionItemKind::STRUCT,
+                    Some(typ.to_string()),
+                ));
+            }
+        }
+    }
+}
 
 pub(super) fn builtin_integer_types() -> [&'static str; 8] {
     ["i8", "i16", "i32", "i64", "u8", "u16", "u32", "u64"]
